+++ internal/dialect/dialectquery/postgres.go
@@ -5,13 +5,6 @@ import (
 	"strings"
 )
 
-const (
-	// defaultSchemaName is the default schema name for Postgres.
-	//
-	// https://www.postgresql.org/docs/current/ddl-schemas.html#DDL-SCHEMAS-PUBLIC
-	defaultSchemaName = "public"
-)
-
 type Postgres struct{}
 
 var _ Querier = (*Postgres)(nil)
@@ -53,14 +46,18 @@ func (p *Postgres) GetLatestVersion(tableName string) string {
 
 func (p *Postgres) TableExists(tableName string) string {
 	schemaName, tableName := parseTableIdentifier(tableName)
-	q := `SELECT EXISTS ( SELECT FROM pg_tables WHERE schemaname = '%s' AND tablename = '%s' )`
-	return fmt.Sprintf(q, schemaName, tableName)
+	if schemaName != "" {
+		q := `SELECT EXISTS ( SELECT 1 FROM pg_tables WHERE schemaname = '%s' AND tablename = '%s' )`
+		return fmt.Sprintf(q, schemaName, tableName)
+	}
+	q := `SELECT EXISTS ( SELECT 1 FROM pg_tables WHERE tablename = '%s' )`
+	return fmt.Sprintf(q, tableName)
 }
 
 func parseTableIdentifier(name string) (schema, table string) {
 	schema, table, found := strings.Cut(name, ".")
 	if !found {
-		return defaultSchemaName, name
+		return "", name
 	}
 	return schema, table
 }