+++ CHANGELOG.md
@@ -7,6 +7,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Fix regression (`v3.23.1` and `v3.24.0`) in postgres migration table existence check for
+  non-default schema. (#882, #883)
+
 ## [v3.24.0]
 
 - Add support for loading environment variables from `.env` files, enabled by default.